requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
google-auth>=2.23.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.100.0
PyGithub>=1.59.0
tenacity>=8.2.0
typing-extensions>=4.7.0 